requirements.txt
lxml==4.9.2
netaddr==0.8.0
PyYAML==6.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
urllib3==1.26.7
zeep==4.2.1